Support nodes with launch condition
Some special nodes when started from an image will not be instantly
available, they might be busy for a while. Nodepool should wait, until
the nodes are ready.

An ssh connection is established to the node and the node is considered
to be ready when a given file has been found on the filesystem of the
node.

Was I67fd9b92e838f8dfc1a4d17bef0390677e36591c (abandoned)

Change-Id: Idc0981c06f9092324362fa571a5ec5deb8bcbd3b