Support install phase with nodepool
Some operating systems are not available as cloud images. To be able to
create nodes of such operating systems, we must install them first on
top of, or by replacing the standard os. This install phase can
include several automated restarts of the node.

To track the installation of the operating system we need more than the
instance's status so an ssh connection is established to check the
existence of a stamp file.

The node preparation script might choose to shut down the node before
the snapshot is taken to guarantee the filesystem's consistency. A step
was added just before snapshot to make sure that the instance is in the
shutoff state before performing the snapshot.