Support nodes with launch condition
Some special nodes when started from an image will not be instantly
available, they might be busy for a while. Nodepool should wait, until
the nodes are ready.

An ssh connection is established to the node and the node is considered
to be ready when a given file has been found on the filesystem of the
node.